Update PLAN.md progress: complete Phase 2.3 (dependency injection), Phase 3.2 (state machine memory management), and Phase 5.2 (MIDI output tests); inline Error/Logger implementations to reduce binary size; wire IMidiOutput through RTPMainUnit→RTPSequencer→RTPScene→sequences; replace BuitStateMachine raw pointers with unique_ptr; add const getSize() overload to RTPScene; migrate RTPThreeAxisVL to HardwareConfig constants

+++ src/Helpers/BuitPersistenceManager.cpp
@@ -174,7 +174,7 @@ bool BuitPersistenceManager::parseAndLoadSequences(RTPSequencer& sequencer, cons
             JsonObject seqObj = sequencesArray[seqIdx];
             
             // Get reference to the sequence
-            RTPEventNoteSequence* sequence = scene->getSequence(seqIdx);
+            RTPEventNoteSequence* sequence = scene->getSequence(static_cast<uint8_t>(seqIdx));
             if (!sequence) continue;
             
             // Load the sequence data
